Expand ALife object IDs from u16 to u32 (65 535 → 4.29B object limit)

+++ src/xrGame/InventoryBox.h
@@ -8,7 +8,7 @@ class CInventoryBox : public CGameObject
 	typedef CGameObject inherited;
 
 public:
-	xr_vector<u16> m_items;
+	xr_vector<u32> m_items;
 
 protected:
 	bool m_in_use;
